+++ tests/solver_tests.rs
@@ -6,7 +6,7 @@ use tsp_solver::utils::compute_circuit_dist;
 
 #[wasm_bindgen_test]
 fn test_greedy_solver(){
-    let mut dist_map:HashMap<(u32,u32),f32> = HashMap::new();
+    let mut dist_map:HashMap<(u8,u8),f32> = HashMap::new();
     dist_map.insert((0,1),1.0);
     dist_map.insert((0,2),1.50);
     dist_map.insert((0,3),1.0);
@@ -14,7 +14,7 @@ fn test_greedy_solver(){
     dist_map.insert((1,3),1.5);
     dist_map.insert((2,3),1.0);
 
-    let mut solver = Solver::new(4,dist_map.clone());
+    let mut solver = Solver::new(4 as u8,dist_map.clone());
     solver.greedy_update();
     assert_eq!(compute_circuit_dist(&dist_map,&solver.route),4.0);
     solver.greedy_update();
@@ -28,7 +28,7 @@ fn test_greedy_solver(){
 
 #[wasm_bindgen_test]
 fn test_ant_solver(){
-    let mut dist_map:HashMap<(u32,u32),f32> = HashMap::new();
+    let mut dist_map:HashMap<(u8,u8),f32> = HashMap::new();
     dist_map.insert((0,1),1.0);
     dist_map.insert((0,2),300.);
     dist_map.insert((0,3),1.0);
@@ -48,7 +48,7 @@ fn test_ant_solver(){
 Test Needs to be deterministic before it can be used as a standard test
 #[wasm_bindgen_test]
 fn test_ant_solver(){
-    let mut dist_map:HashMap<(u32,u32),f32> = HashMap::new();
+    let mut dist_map:HashMap<(u8,u8),f32> = HashMap::new();
     dist_map.insert((0,1),1.0);
     dist_map.insert((0,2),300.);
     dist_map.insert((0,3),1.0);
@@ -56,7 +56,7 @@ fn test_ant_solver(){
     dist_map.insert((1,3),300.);
     dist_map.insert((2,3),1.0);
 
-    let mut pher_map:HashMap<(u32,u32),f32> = HashMap::new();
+    let mut pher_map:HashMap<(u8,u8),f32> = HashMap::new();
     pher_map.insert((0,1),1.0);
     pher_map.insert((0,2),5.0);
     pher_map.insert((0,3),1.0);
@@ -73,7 +73,7 @@ fn test_ant_solver(){
     assert_eq!(compute_circuit_dist(&dist_map,&solver.route),4.0);
 
 
-    let mut dist_map:HashMap<(u32,u32),f32> = HashMap::new();
+    let mut dist_map:HashMap<(u8,u8),f32> = HashMap::new();
     dist_map.insert((0,1),2.0);
     dist_map.insert((0,2),3.);
     dist_map.insert((0,3),2.0);
@@ -81,7 +81,7 @@ fn test_ant_solver(){
     dist_map.insert((1,3),3.);
     dist_map.insert((2,3),2.0);
 
-    let mut pher_map:HashMap<(u32,u32),f32> = HashMap::new();
+    let mut pher_map:HashMap<(u8,u8),f32> = HashMap::new();
     pher_map.insert((0,1),1.0);
     pher_map.insert((0,2),300.0);
     pher_map.insert((0,3),1.0);
